--- src/main/java/ByeCommand.java
@@ -1,11 +1,0 @@
-public class ByeCommand extends Command{
-
-    public void execute(TaskList list, Ui ui, Storage storage) {
-
-    }
-
-    public boolean isExit() {
-        return true;
-    }
-
-}
