--- src/main/java/ShowListCommand.java
@@ -1,13 +1,0 @@
-public class ShowListCommand extends Command{
-
-
-
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.listUI(tasks);
-    }
-
-    public boolean isExit() {
-        return false;
-    }
-
-}
