+++ src/main/java/duke/Commands.java
@@ -1,3 +1,5 @@
+package duke;
+
 public enum Commands {
     UNMARK, MARK, TODO, DEADLINE, EVENT, DELETE, LIST, BYE
 }
