+++ src/main/java/duke/command/Command.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.*;
+import duke.exceptions.DirectoryNotFoundException;
+import duke.exceptions.DukeException;
+import duke.storage.Storage;
+import duke.ui.Ui;
+
 import java.io.FileNotFoundException;
 import java.io.IOException;
