+++ src/main/java/duke/command/MarkDoneCommand.java
@@ -1,7 +1,17 @@
+package duke.command;
+
+import duke.TaskList;
+import duke.command.Command;
+import duke.exceptions.DirectoryNotFoundException;
+import duke.exceptions.DukeException;
+import duke.storage.Storage;
+import duke.tasks.Task;
+import duke.ui.Ui;
+
 import java.io.FileNotFoundException;
 import java.io.IOException;
 
-public class MarkDoneCommand extends Command{
+public class MarkDoneCommand extends Command {
 
     protected int index;
