+++ src/main/java/duke/exceptions/DirectoryNotFoundException.java
@@ -1,7 +1,9 @@
+package duke.exceptions;
+
 public class DirectoryNotFoundException extends Exception{
 
 
-    DirectoryNotFoundException() {
+    public DirectoryNotFoundException() {
 
     }
