+++ src/main/java/duke/exceptions/DukeException.java
@@ -1,6 +1,8 @@
+package duke.exceptions;
+
 public class DukeException extends Exception{
     String message;
-    DukeException(String str) {
+    public DukeException(String str) {
         message = str;
     }
     public String toString() {
