+++ src/main/java/duke/exceptions/ToDoException.java
@@ -1,6 +1,8 @@
+package duke.exceptions;
+
 public class ToDoException extends Exception{
     String message;
-    ToDoException(String str) {
+    public ToDoException(String str) {
         message = str;
     }
     public String toString() {
