+++ src/main/java/duke/tasks/Deadline.java
@@ -1,8 +1,9 @@
+package duke.tasks;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.time.temporal.ChronoUnit;
 
-public class Deadline extends Task{
+public class Deadline extends Task {
 
     protected String eventDetail;
