+++ src/main/java/duke/tasks/Todo.java
@@ -1,4 +1,6 @@
-public class Todo extends Task{
+package duke.tasks;
+
+public class Todo extends Task {
 
     public String todo;
