+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ------------------------------------------------------------------------------
-Hello! I'm Duke
+Hello! I'm duke.Duke
 What can I do for you?
 ------------------------------------------------------------------------------
 ------------------------------------------------------------------------------
